v0.4
Re-insulated the sodded wires for the IR-sensor, sodded longer wires onto the legs of the thermistor, to extend its' reach.
Implemented the buzzer to which there was already code.

Tweaked some code.

Added new code for a better way of figuring out whether the stove was out of control and a better alarming system.

Re-organized the wires and added another breadboard.

Co-Authored-By: kiwra <kiwra@users.noreply.github.com>
Co-Authored-By: sivnus <sivnus@users.noreply.github.com>